+++ src/main/java/com/influy/domain/item/controller/ItemRestController.java
@@ -135,35 +135,4 @@ public class ItemRestController {
     public ApiResponse<ItemResponseDto.TalkBoxOpenedListDto> getTalkBoxOpened (@AuthenticationPrincipal CustomUserDetails userDetails) {
         return ApiResponse.onSuccess(itemService.getTalkBoxOpened(userDetails));
     }
-
-    @GetMapping("/home/close-deadline")
-    @Operation(summary = "홈 마감 임박 상품 페이지", description = "24시간 이내 마감 상품")
-    public ApiResponse<ItemResponseDto.HomeItemViewPageDto> getCloseDeadline(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                                                             @Valid @ParameterObject PageRequestDto pageRequest) {
-        return ApiResponse.onSuccess(itemService.getCloseDeadline(userDetails, pageRequest));
-    }
-
-    @GetMapping("/home/popular")
-    @Operation(summary = "홈 인기 급상승 상품 페이지", description = "3개 반환 (나중에 더보기 생길 것을 고려해 pageable 사용)")
-    public ApiResponse<ItemResponseDto.HomeItemViewPageDto> getPopular(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                                                             @Valid @ParameterObject PageRequestDto pageRequest) {
-        return ApiResponse.onSuccess(itemService.getPopular(userDetails, pageRequest));
-    }
-
-    @GetMapping("/home/recommend")
-    @Operation(summary = "홈 추천 상품 페이지", description = "itemCategory request param 없이 요청보내면 '전체' 카테고리 결과 반환")
-    public ApiResponse<ItemResponseDto.HomeItemViewPageDto> getRecommended(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                                                             @Valid @ParameterObject PageRequestDto pageRequest,
-                                                                               @RequestParam(name = "categoryId", required = false) Long categoryId) {
-        return ApiResponse.onSuccess(itemService.getRecommended(userDetails, pageRequest, categoryId));
-    }
-
-    @GetMapping("seller/home/questions")
-    @Operation(summary = "셀러 홈 상품 질문 안내", description = "아이템 + 아이템 별 많이 들어오는 질문 카테고리")
-    public ApiResponse<ItemResponseDto.SellerHomeItemPageDTO> getSellerHomeItem(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                                                            @Valid @ParameterObject PageRequestDto pageRequestDto){
-        SellerProfile seller = memberService.checkSeller(userDetails);
-        return ApiResponse.onSuccess(itemService.getSellerHomeItemWithQuestionStatus(seller, pageRequestDto));
-    }
-
 }
