+++ src/main/java/com/influy/domain/item/converter/ItemConverter.java
@@ -182,70 +182,4 @@ public class ItemConverter {
                 .talkBoxOpenedDtoList(itemDtoList)
                 .build();
     }
-
-    public static ItemResponseDto.HomeItemViewDto toHomeItemViewDto(Item item, Boolean liked) {
-        Member member = item.getSeller().getMember();
-
-        return ItemResponseDto.HomeItemViewDto.builder()
-                .sellerNickname(member.getNickname())
-                .sellerId(item.getSeller().getId())
-                .sellerProfileImg(member.getProfileImg())
-                .sellerUsername(member.getUsername())
-                .itemId(item.getId())
-                .itemName(item.getName())
-                .itemMainImg(item.getMainImg())
-                .itemPeriod(item.getItemPeriod())
-                .startDate(item.getStartDate())
-                .endDate(item.getEndDate())
-                .tagline(item.getTagline())
-                .currentStatus(item.getItemStatus())
-                .liked(liked)
-                .build();
-    }
-
-    public static ItemResponseDto.HomeItemViewPageDto toHomeItemViewPageDto(Page<Item> itemPage, List<Long> likeItems) {
-        List<Long> safeLikeItems = (likeItems != null) ? likeItems : Collections.emptyList();
-
-        List<ItemResponseDto.HomeItemViewDto> itemDtoList = itemPage != null
-                ? itemPage.stream()
-                .map(item -> toHomeItemViewDto(item, safeLikeItems.contains(item.getId())))
-                .toList()
-                : Collections.emptyList();
-
-        return ItemResponseDto.HomeItemViewPageDto.builder()
-                .itemPreviewList(itemDtoList)
-                .listSize(itemDtoList.size())
-                .totalPage(itemPage == null ? 0: itemPage.getTotalPages())
-                .totalElements(itemPage == null ? 0: itemPage.getTotalElements())
-                .isFirst(itemPage == null || itemPage.isFirst())
-                .isLast(itemPage == null || itemPage.isLast())
-                .build();
-    }
-
-    public static ItemResponseDto.SellerHomeItemDTO toSellerHomeItemDTO(ItemJPQLResponse.ItemWithQuestionStatus itemJPQLResult, List<String> topNCategories) {
-        return ItemResponseDto.SellerHomeItemDTO.builder()
-                .itemId(itemJPQLResult.getItem().getId())
-                .imageUrl(itemJPQLResult.getItem().getMainImg())
-                .itemTitle(itemJPQLResult.getItem().getName())
-                .itemStatus(itemJPQLResult.getItem().getItemStatus())
-                .itemPeriod(itemJPQLResult.getItem().getItemPeriod())
-                .startDate(itemJPQLResult.getItem().getStartDate())
-                .endDate(itemJPQLResult.getItem().getEndDate())
-                .newQuestions(itemJPQLResult.getNewQuestions())
-                .totalPendingQuestions(itemJPQLResult.getPendingQuestions())
-                .topCategories(topNCategories)
-                .build();
-    }
-
-    public static ItemResponseDto.SellerHomeItemPageDTO toSellerHomeItemPageDTO(Page<ItemJPQLResponse.ItemWithQuestionStatus> itmePage, Map<Long, List<String>> topNCategoryMap) {
-        List<ItemResponseDto.SellerHomeItemDTO> itemsList = itmePage.stream().map(item->toSellerHomeItemDTO(item, topNCategoryMap.get(item.getItem().getId()))).toList();
-        return ItemResponseDto.SellerHomeItemPageDTO.builder()
-                .itemList(itemsList)
-                .totalPage(itmePage.getTotalPages())
-                .totalElements(itmePage.getTotalElements())
-                .isFirst(itmePage.isFirst())
-                .isLast(itmePage.isLast())
-                .listSize(itemsList.size())
-                .build();
-    }
 }
