+++ src/main/java/com/influy/domain/search/service/SearchService.java
@@ -1,5 +1,6 @@
 package com.influy.domain.search.service;
 
+import com.influy.domain.home.dto.HomeResponseDto;
 import com.influy.domain.item.dto.ItemResponseDto;
 import com.influy.domain.search.dto.SearchResponseDto;
 import com.influy.global.common.PageRequestDto;
@@ -7,5 +8,5 @@ import com.influy.global.jwt.CustomUserDetails;
 
 public interface SearchService {
     SearchResponseDto.SellerPageResultDto searchSeller(CustomUserDetails userDetails, String query, PageRequestDto pageRequest);
-    ItemResponseDto.HomeItemViewPageDto searchItem(CustomUserDetails userDetails, String query, PageRequestDto pageRequest);
+    HomeResponseDto.HomeItemViewPageDto searchItem(CustomUserDetails userDetails, String query, PageRequestDto pageRequest);
 }
