+++ src/main/java/com/influy/domain/sellerProfile/repository/SellerProfileRepository.java
@@ -1,5 +1,6 @@
 package com.influy.domain.sellerProfile.repository;
 
+import com.influy.domain.item.dto.ItemResponseDto;
 import com.influy.domain.sellerProfile.entity.SellerProfile;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -7,6 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface SellerProfileRepository extends JpaRepository<SellerProfile, Long> {
@@ -26,4 +28,7 @@ public interface SellerProfileRepository extends JpaRepository<SellerProfile, Lo
 
     @Query("SELECT EXISTS (SELECT 1 FROM QuestionCategory c WHERE c.id = :categoryId AND c.item.seller.id = :sellerId)")
     Boolean existsByIdAndCategoryId(@Param("sellerId") Long sellerId, @Param("categoryId") Long categoryId);
+
+
+    List<SellerProfile> findTop10ByIsPublicTrue();
 }
